Adds rq for workers, supports internal / external
