+++ server/settings.py
@@ -21,3 +21,4 @@ SLACK_TOKEN = os.getenv(
 OPENAI_API_KEY = os.getenv(
     "OPENAI_API_KEY", "sk-1234567890-1234567890-1234567890-1234567890"
 )
+EXTERNAL_WORKERS = os.getenv("EXTERNAL_WORKERS", False)
